# api-docs/openapi.yaml
openapi: 3.0.0
info:
  title: ModelLab API
  description: |
    ModelLab is a comprehensive ML experiment tracking platform.

    This API provides endpoints for managing ML projects, datasets, training runs, and artifacts.

    **Key Features:**
    - Project workspace organization
    - Dataset upload and management
    - Training run tracking with metrics and hyperparameters
    - Reproducibility packs for experiment reproduction
    - Comprehensive model evaluation with EvalHarness

    **Base URL:** `http://localhost:3001/api/modellab`

    **Authentication:** Not yet implemented (Phase 2)

  version: 1.0.0
  contact:
    name: ModelLab Support
    url: https://github.com/calebnewtonusc/ModelLab
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api/modellab
    description: Local development server
  - url: https://modellab.studio/api/modellab
    description: Production server
...
paths:
...
  /projects/{id}/datasets:
    get:
      tags:
        - Projects
      summary: List datasets in a project
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  datasets:
                    type: array
                    items:
                      $ref: '#/components/schemas/Dataset'
...
  /datasets:
    get:
      tags:
        - Datasets
      summary: List all datasets
      parameters:
        - name: project_id
          in: query
          schema:
            type: string
          description: Filter by project ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  datasets:
                    type: array
                    items:
                      $ref: '#/components/schemas/Dataset'
...
  /datasets/{id}:
    get:
      tags:
        - Datasets
      summary: Get a specific dataset
      parameters:
        - $ref: '#/components/parameters/DatasetId'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  dataset:
                    $ref: '#/components/schemas/Dataset'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Datasets
      summary: Update dataset metadata
      parameters:
        - $ref: '#/components/parameters/DatasetId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Dataset updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  dataset:
                    $ref: '#/components/schemas/Dataset'

    delete:
      tags:
        - Datasets
      summary: Delete a dataset
      parameters:
        - $ref: '#/components/parameters/DatasetId'
      responses:
        '204':
          description: Dataset deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
  /datasets/{id}/preview:
    get:
      tags:
        - Datasets
      summary: Preview dataset rows
      parameters:
        - $ref: '#/components/parameters/DatasetId'
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 1000
          description: Number of rows to return
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  rows:
                    type: array
                    items:
                      type: object

  # ==================== Runs ====================
...
components:
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Project ID
    DatasetId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Dataset ID
...
  schemas:
...
    Dataset:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        project_id:
          type: string
          nullable: true
        filePath:
          type: string
        fileSize:
          type: integer
        rowCount:
          type: integer
        columns:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Resource not found"
...